consolidated rendering into single file
