--- demo/tc-ui/lib/render.jsx
@@ -1,13 +1,0 @@
-const { render } = require('preact-render-to-string');
-const { h } = require('preact');
-const Layout = require('../templates/layout/layout');
-
-module.exports = (Template, props) => {
-	return `
-<!DOCTYPE html>
-${render(
-	<Layout {...props}>
-		<Template {...props} />
-	</Layout>,
-)}`;
-};
